feat: refactor project structure
