+++ package.json
@@ -6,14 +6,14 @@
     "create-db": "netlify dev:exec node scripts/create-database.js",
     "create-data": "netlify dev:exec node scripts/create-example-data.js",
     "bootstrap": "npx npm-run-all create-db create-data",
-    "watch:src": "parcel --log-level 2 src/index.html",
-    "watch:functions": "netlify-lambda serve functions -c webpack.dev.js",
-    "dev": "npx npm-run-all -p watch:src watch:functions",
-    "prebuild": "npm run format && npm run build:functions",
-    "build:functions": "netlify-lambda build functions -c webpack.prod.js",
-    "build": "parcel build src/index.html",
-    "serve": "parcel serve src/index.html",
-    "format": "prettier \"{functions,scripts,src}/**/*.{js,ts,tsx}\" --write",
+    "watch:src": "parcel --log-level 2 src/client/ui/index.html",
+    "watch:server": "netlify-lambda serve src/server -c webpack.dev.js",
+    "dev": "npx npm-run-all -p watch:src watch:server",
+    "prebuild": "npm run format && npm run build:server",
+    "build:server": "netlify-lambda build src/server -c webpack.prod.js",
+    "build": "parcel build src/client/ui/index.html",
+    "serve": "parcel serve src/client/ui/index.html",
+    "format": "prettier \"{server,scripts,src}/**/*.{js,ts,tsx}\" --write",
     "lint": "eslint \"src/**/*.{js,jsx}\" --quiet"
   },
   "keywords": [
@@ -46,7 +46,6 @@
     "@types/node": "^14.14.22",
     "apollo-server-testing": "^2.19.2",
     "ava": "^3.15.0",
-    "babel-plugin-inline-react-svg": "^1.1.2",
     "eslint": "^7.14.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-import": "^2.22.1",
