+++ src/client/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Render React application.
+ */
+import './ui'
