+++ src/client/ui/assets/index.ts
