+++ src/client/ui/index.tsx
@@ -8,24 +8,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
 
-import {
-  ApolloClient,
-  HttpLink,
-  InMemoryCache,
-  ApolloProvider,
-} from '@apollo/client'
-
-const cache = new InMemoryCache()
-
-const link = new HttpLink({
-  uri: '/.netlify/functions/graphql',
-})
-
-const client = new ApolloClient({
-  cache,
-  link,
-  // resolvers
-})
+import { client } from '../graphql'
+import { ApolloProvider } from '@apollo/client'
 
 const Root = () => (
   <ApolloProvider client={client}>
