+++ src/server/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Expose graphQL logic.
+ */
+export * from './graphql'
