+++ tsconfig.json
@@ -10,6 +10,6 @@
   "include": [
     "**/*.ts",
     "**/*.tsx",
-    "src/assets/**",
+    "src/**/*"
   ]
 }
